Feature/25 member invitation

+++ src/main/java/com/example/splitly/application/service/GroupInfoImpl.java
@@ -1,17 +1,22 @@
 package com.example.splitly.application.service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-
 
 import org.springframework.stereotype.Service;
 
 import com.example.splitly.domain.repository.UserRepository;
 import com.example.splitly.presentation.dto.request.GroupDTO;
 import com.example.splitly.application.serviceInterface.IGroupInfo;
+import com.example.splitly.application.serviceInterface.IGroupUser;
 import com.example.splitly.domain.entity.GroupInfo;
+import com.example.splitly.domain.entity.GroupUser;
+import com.example.splitly.domain.entity.GroupUserId;
 import com.example.splitly.domain.entity.User;
+import com.example.splitly.domain.enumerator.InvitationStatus;
 import com.example.splitly.domain.repository.GroupInfoRepository;
+import com.example.splitly.domain.repository.GroupUserRepository;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -22,6 +27,8 @@ public class GroupInfoImpl implements IGroupInfo {
 
     private final GroupInfoRepository groupInfoRepository;
     private final UserRepository userRepository;
+    private final IGroupUser iGroupUser;
+    private final GroupUserRepository groupUserRepository;
 
     @Override
     public List<GroupInfo> getAllGroup() {
@@ -31,29 +38,62 @@ public class GroupInfoImpl implements IGroupInfo {
     }
 
     @Override
-    public GroupInfo createGroup(GroupDTO groupDTO) {
+    public GroupInfo createGroup(GroupDTO groupDTO, Integer userId, List<String> emailList) {
         GroupInfo groupInfo = GroupInfo.builder()
                 .numberOfMember(groupDTO.getNumberOfMember())
                 .groupName(groupDTO.getGroupName())
                 .build();
         groupInfoRepository.save(groupInfo);
+        assignLeader(groupInfo.getGroupId(), userId);
+        GroupUserId groupUserId = GroupUserId.builder()
+                        .groupId(groupInfo.getGroupId())
+                        .userId(userId)
+                        .build();
+        GroupUser groupUser = GroupUser.builder()
+                        .groupUserId(groupUserId)
+                        .groupInfo(groupInfo)
+                        .user(groupInfo.getUser())
+                        .joinedAt(LocalDateTime.now())
+                        .status(InvitationStatus.SUCCESS)
+                    .build();
+        groupUserRepository.save(groupUser);
+        if (!emailList.isEmpty()) {
+            for (String email : emailList) {
+                try {
+                    if (email != groupInfo.getUser().getEmail()) {
+                        iGroupUser.inviteUserToGroup(email, groupInfo.getGroupId());
+                    }
+                } catch (EntityNotFoundException e) {
+                    System.out.println("Email not found: " + email);
+                }
+            }
+        }
         return groupInfo;
     }
 
     @Override
-    public GroupInfo updateGroup(Long groupId, GroupDTO groupDTO) {
+    public GroupInfo updateGroup(Long groupId, GroupDTO groupDTO, Integer leaderId) {
         GroupInfo groupInfo = findGroupInfo(groupId);
-        groupInfo.setGroupName(groupDTO.getGroupName());
-        groupInfo.setNumberOfMember(groupDTO.getNumberOfMember());
-        groupInfoRepository.save((groupInfo));
+        if (groupInfo.getUser().getUserId() == leaderId) {
+            groupInfo.setGroupName(groupDTO.getGroupName());
+            groupInfo.setNumberOfMember(groupDTO.getNumberOfMember());
+            groupInfoRepository.save((groupInfo));
+        } else {
+            throw new EntityNotFoundException("User is not a leader");
+        }
         return groupInfo;
     }
 
     @Override
-    public void deleteGroup(Long groupId) {
+    public void deleteGroup(Long groupId, Integer leaderId) {
+        iGroupUser.deleteGroupUser(groupId);
         GroupInfo groupInfo = findGroupInfo(groupId);
-        if (groupInfo != null) {
-            groupInfoRepository.deleteById(groupId);
+        if (groupInfo != null && groupInfo.getUser().getUserId() == leaderId) {
+            groupInfo.setUser(null);
+            groupInfoRepository.save(groupInfo);
+            groupInfoRepository.deleteById(groupId); 
+        } else {
+            throw new EntityNotFoundException("Group not found!");
         }
     }
